+++ src/lib/ReactDataTable/TableRows.tsx
@@ -3,29 +3,38 @@ import { Row, flexRender } from "@tanstack/react-table";
 import { CSSProperties } from "react";
 import { CSS } from "@dnd-kit/utilities";
 import { getCommonPinningStyles } from "../utils/getCommonPinningStyles";
+import { FilterModel } from "../types/TableState";
+import { ReactDataTableProps } from "./ReactDataTableProps";
 
-interface TableRowProps<TData> {
+interface TableRowProps<TData, TFilter extends FilterModel = Record<string, never>>
+  extends Pick<ReactDataTableProps<TData, TFilter>, "onRowClick" | "enableRowClick"> {
   row: Row<TData>;
   enableRowSelection?: boolean | ((row: Row<TData>) => boolean);
+  fullRowSelectable?: boolean;
   enableExpanding?: boolean | ((row: Row<TData>) => boolean);
   rowStyle?: CSSProperties;
   setNodeRef?: (node: HTMLElement | null) => void;
 }
 
-const InternalTableRow = <TData,>(props: TableRowProps<TData>) => {
-  const { row, rowStyle, setNodeRef, enableRowSelection = false } = props;
-  const isRowSelectionEnabled = typeof enableRowSelection === "function" ? enableRowSelection(row) : enableRowSelection;
-
+const InternalTableRow = <TData, TFilter extends FilterModel = Record<string, never>>(props: TableRowProps<TData, TFilter>) => {
+  const { row, rowStyle, setNodeRef, enableRowSelection = false, fullRowSelectable = true, onRowClick, enableRowClick } = props;
+  const isRowSelectionEnabled =
+    (typeof enableRowSelection === "function" ? enableRowSelection(row) : enableRowSelection) && fullRowSelectable;
+  const isRowClickable = typeof enableRowClick === "function" ? enableRowClick(row) : enableRowClick;
   return (
     <tr
       key={row.id}
       ref={setNodeRef}
-      onClick={() => {
+      onClick={async () => {
         if (isRowSelectionEnabled) {
           row.toggleSelected();
+        } else if (isRowClickable && onRowClick) {
+          await onRowClick(row);
+        } else {
+          // Nothing to execute
         }
       }}
-      className={isRowSelectionEnabled ? "cursor-pointer" : undefined}
+      className={isRowSelectionEnabled || isRowClickable ? "cursor-pointer" : undefined}
       style={rowStyle}
     >
       {row.getVisibleCells().map((cell) => (
@@ -41,8 +50,8 @@ const InternalTableRow = <TData,>(props: TableRowProps<TData>) => {
   );
 };
 
-const DraggableRow = <TData,>(props: TableRowProps<TData>) => {
-  const { row, rowStyle, enableRowSelection, enableExpanding } = props;
+const DraggableRow = <TData, TFilter extends FilterModel = Record<string, never>>(props: TableRowProps<TData, TFilter>) => {
+  const { row, rowStyle } = props;
   const { transform, transition, setNodeRef, isDragging } = useSortable({ id: row.id });
   const draggableStyle: CSSProperties = {
     transform: CSS.Transform.toString(transform),
@@ -52,10 +61,8 @@ const DraggableRow = <TData,>(props: TableRowProps<TData>) => {
     position: "relative",
   };
   return (
-    <InternalTableRow
-      row={row}
-      enableRowSelection={enableRowSelection}
-      enableExpanding={enableExpanding}
+    <InternalTableRow<TData, TFilter>
+      {...props}
       setNodeRef={setNodeRef}
       rowStyle={rowStyle ? { ...rowStyle, ...draggableStyle } : draggableStyle}
     />